Remove permute.py
Its job is entirely handled with Qobj now, except for `reshuffle` (I'm
not even entirely sure what it does with that name...) which is purely a
superoperator thing, so is moved there.
