+++ qutip/core/superoperator.py
@@ -399,7 +399,6 @@ def sprepost(A, B):
     """
     if isinstance(A, QobjEvo) or isinstance(B, QobjEvo):
         return spre(A) * spost(B)
-
     else:
         dims = [[_drop_projected_dims(A.dims[0]),
                  _drop_projected_dims(B.dims[1])],
@@ -407,5 +406,31 @@ def sprepost(A, B):
                  _drop_projected_dims(B.dims[0])]]
         data = zcsr_kron(B.data.T, A.data)
         return Qobj(data, dims=dims, superrep='super')
+
+
+def reshuffle(q_oper):
+    """
+    Column-reshuffles a ``type="super"`` Qobj.
+    """
+    if q_oper.type not in ('super', 'operator-ket'):
+        raise TypeError("Reshuffling is only supported on type='super' "
+                        "or type='operator-ket'.")
+    # How many indices are there, and how many subsystems can we decompose
+    # each index into?
+    n_indices = len(q_oper.dims[0])
+    n_subsystems = len(q_oper.dims[0][0])
+    # Generate a list of lists (lol) that represents the permutation order we
+    # need. It's easiest to do so if we make an array, then turn it into a lol
+    # by using map(list, ...). That array is generated by using reshape and
+    # transpose to turn an array like [a, b, a, b, ..., a, b] into one like
+    # [a, a, ..., a, b, b, ..., b].
+    perm_idxs = map(list,
+                    np.arange(n_subsystems * n_indices)[
+                        np.arange(n_subsystems * n_indices).reshape(
+                            (n_indices, n_subsystems)).T.flatten()
+                    ].reshape((n_subsystems, n_indices))
+                    )
+    return q_oper.permute(list(perm_idxs))
+
 
 from .qobjevo import QobjEvo
